Remove `is_mixer_irq_important()` stuff
The idea behind this was that when accurate timing is "not important",
feeding the mix buffer with samples be can be done in a "less accurate"
manner by performing some stretching on the samples to avoid clicks and
pops.

However, in actual practice this does not improve anything in moderate
to severe buffer under/overflow situations, so it's just a complication
that has to go.